+++ vite.doc.config.js
@@ -2,6 +2,7 @@ import vue from '@vitejs/plugin-vue';
 const path = require('path');
 const fs = require('fs');
 const hljs = require('highlight.js/lib/common');
+const marked = require('marked');
 
 const sampleCodePlugin = {
   name: 'vue-sample-code',
@@ -20,10 +21,16 @@ const sampleCodePlugin = {
           `<pre><code class="language-html hljs">${content}</code></pre>`
         );
       });
-      if (result.length) {
-        return { code, map: null };
-      }
+      const changelogs = [...code.matchAll(/\<change-log \/\>/g)];
+      changelogs.forEach((item) => {
+        let content = fs
+          .readFileSync(path.resolve(__dirname, 'CHANGELOG.md'))
+          .toString();
+        content = '<article class="markdown-body">' + marked.parse(content) + '</article>';
+        code = code.replace(item[0], content);
+      });
     }
+    return { code, map: null };
   },
 };